@@ -33,6 +33,7 @@
     "cookie-parser": "^1.4.3",
     "cookie-sessions": "github:auth0/cookie-sessions#53a8aae",
     "csurf": "1.9.0",
+    "diff_match_patch": "^0.1.1",
     "ejs": "^2.5.5",
     "express": "^4.16.4",
     "express-passport-logout": "~0.1.0",
@@ -74,6 +75,6 @@
   },
   "optionalDependencies": {},
   "engines": {
-    "node": "^10.16"
+    "node": "^12.19"
   }
 }
