@@ -4,7 +4,7 @@ const binary = require('binary');
 // https://github.com/pfmooney/node-ldapjs-mangle-proxy/blob/master/lib/objectsid.js
 
 function outputType(type) {
-  var buf = new Buffer(6);
+  var buf = Buffer.alloc(6);
   type.copy(buf, 0);
   var output = 0;
   for (var i = 0; i < 6; i++) {
